fixup! fixup! fixup! fixup! feat: wire up initial device class and zones
